Move test to src/lib.rs
